New method names, clarification on handshake
Fixes the new readme example to use the new method names that are coming
in RC2.

Adds clarification to the docs for Websocket::handleHandshake to clarify
that changing the status code rejects the connection as opposed to
returning a new Response object.

Fixes the PHP requirement specified in the readme.